+++ src/document/plugin.ts
@@ -9,7 +9,7 @@ import {
   WidgetTracker
 } from '@jupyterlab/apputils';
 import { IEditorServices } from '@jupyterlab/codeeditor';
-import { PageConfig, PathExt, URLExt } from '@jupyterlab/coreutils';
+import { PageConfig, URLExt } from '@jupyterlab/coreutils';
 import { IDocumentManager } from '@jupyterlab/docmanager';
 import { IDefaultFileBrowser } from '@jupyterlab/filebrowser';
 import { INotebookTracker, NotebookPanel } from '@jupyterlab/notebook';
@@ -33,7 +33,8 @@ import {
   isSpectaApp,
   readSpectaConfig,
   registerDocumentFactory,
-  PLAINB_PREFIX
+  getSpectaDocInfo,
+  openDocument
 } from '../tool';
 
 const activate = (
@@ -141,16 +142,9 @@ export const spectaOpener: JupyterFrontEndPlugin<void, ILabShell> = {
       app.restored.then(async () => {
         const labShell = app.shell;
 
-        const fileTypes = app.docRegistry.getFileTypesForPath(path);
-        const plainbFileType = fileTypes.find(ft =>
-          ft.name.startsWith(PLAINB_PREFIX)
-        );
-        const isPlainb = !!plainbFileType;
-        const fileTypeName = plainbFileType
-          ? plainbFileType.name
-          : fileTypes[0]?.name;
+        const { isSpectaDoc, factory } = getSpectaDocInfo(path, app);
 
-        if (PathExt.extname(path) === '.ipynb' || isPlainb) {
+        if (isSpectaDoc) {
           const commands = app.commands;
           const spectaConfig = readSpectaConfig({});
           await configLabLayout({
@@ -158,14 +152,7 @@ export const spectaOpener: JupyterFrontEndPlugin<void, ILabShell> = {
             labShell,
             commands
           });
-          let factory = 'specta';
-          if (isPlainb) {
-            factory = fileTypeName.replace(PLAINB_PREFIX, 'specta-');
-          }
-          const widget = docManager.openOrReveal(path, factory);
-          if (widget) {
-            app.shell.add(widget, 'main');
-          }
+          openDocument(path, factory, docManager, app.shell);
         }
       });
       return;
@@ -182,25 +169,11 @@ export const spectaOpener: JupyterFrontEndPlugin<void, ILabShell> = {
         app.restored.then(async () => {
           await new Promise(r => setTimeout(r, 100));
           await kernelSpecManager.ready;
-          const fileTypes = app.docRegistry.getFileTypesForPath(path);
-          const plainbFileType = fileTypes.find(ft =>
-            ft.name.startsWith(PLAINB_PREFIX)
-          );
-          const isPlainb = !!plainbFileType;
-          const fileTypeName = plainbFileType
-            ? plainbFileType.name
-            : fileTypes[0]?.name;
+          const { isSpectaDoc, factory } = getSpectaDocInfo(path, app);
 
-          if (PathExt.extname(path) === '.ipynb' || isPlainb) {
+          if (isSpectaDoc) {
             app.shell.addClass('specta-document-viewer');
-            let factory = 'specta';
-            if (isPlainb) {
-              factory = fileTypeName.replace(PLAINB_PREFIX, 'specta-');
-            }
-            const widget = docManager.openOrReveal(path, factory);
-            if (widget) {
-              app.shell.add(widget, 'main');
-            }
+            openDocument(path, factory, docManager, app.shell);
           } else {
             let count = 0;
             const tryOpen = () => {
